combine two tests into one
